Moved to metadata package and added railway
