+++ java/src/main/java/net/danielgill/railostools/railway/Railway.java
@@ -1,0 +1,5 @@
+package net.danielgill.railostools.railway;
+
+public class Railway {
+    
+}
