+++ java/src/test/java/net/danielgill/railostools/metadata/MetadataTest.java
@@ -1,4 +1,4 @@
-package net.danielgill.railostools;
+package net.danielgill.railostools.metadata;
 
 import java.io.File;
